feat(ios): unify conversation and engine rendering
- unify Message model - merge EngineMessage into Message
- unify status bar - extend ConversationStatusBar, delete EngineFooterView
- unify conversation rendering - adopt EngineMessageRow, delete MessageBubble
- remove bubble wrapper from assistant messages in conversation mode
- add mic button to engine view input bar
- route compaction messages to CompactionRowView in EngineView
